Move download module to utilities
So that the download folder only contains data-set
specific modules while all generalisable utilities are
kept in utilities
